feat: Context menus for navigator, canvas and tokens (#5511)
## Description

- [x] context menu for navigator
- [x] context menu for canvas instances
- [x] right click to open token menu
- [x] pages and folders context menu
- [x] duplicate pages

## Steps for reproduction

1. click button
2. expect xyz

## Code Review

- [ ] hi @kof, I need you to do
  - conceptual review (architecture, feature-correctness)
  - detailed review (read every line)
  - test it on preview

## Before requesting a review

- [ ] made a self-review
- [ ] added inline comments where things may be not obvious (the "why",
not "what")

## Before merging

- [ ] tested locally and on preview environment (preview dev login:
0000)
- [ ] updated [test
cases](https://github.com/webstudio-is/webstudio/blob/main/apps/builder/docs/test-cases.md)
document
- [ ] added tests
- [ ] if any new env variables are added, added them to `.env` file

+++ apps/builder/app/builder/features/pages/page-settings.tsx
@@ -55,9 +55,7 @@ import {
   Switch,
   TitleSuffixSpacer,
   ProBadge,
-  Dialog,
   DialogClose,
-  DialogContent,
   DialogTitle,
   DialogTitleActions,
 } from "@webstudio-is/design-system";
@@ -69,7 +67,6 @@ import {
   UploadIcon,
 } from "@webstudio-is/icons";
 import { useIds } from "~/shared/form-utils";
-import { updateWebstudioData } from "~/shared/instance-utils";
 import {
   $assets,
   $instances,
@@ -103,7 +100,6 @@ import { SearchPreview } from "./search-preview";
 import { SocialPreview } from "./social-preview";
 import {
   registerFolderChildMutable,
-  deletePageMutable,
   $pageRootScope,
   duplicatePage,
   isPathAvailable,
@@ -1457,7 +1453,7 @@ export const PageSettings = ({
 }: {
   onClose: () => void;
   onDuplicate: (newPageId: string) => void;
-  onDelete: () => void;
+  onDelete?: () => void;
   pageId: string;
 }) => {
   const pages = useStore($pages);
@@ -1466,9 +1462,6 @@ export const PageSettings = ({
   const isHomePage = page?.id === pages?.homePage.id;
 
   const [unsavedValues, setUnsavedValues] = useState<Partial<Values>>({});
-
-  const [showDeleteConfirmation, setShowDeleteConfirmation] =
-    useState<boolean>(false);
 
   const values: Values = {
     ...(page ? toFormValues(page, pages, isHomePage) : fieldDefaultValues),
@@ -1526,14 +1519,9 @@ export const PageSettings = ({
   });
 
   const handleRequestDelete = () => {
-    setShowDeleteConfirmation(true);
-  };
-
-  const hanldeDelete = () => {
-    updateWebstudioData((data) => {
-      deletePageMutable(pageId, data);
-    });
-    onDelete();
+    if (onDelete) {
+      onDelete();
+    }
   };
 
   if (page === undefined) {
@@ -1563,68 +1551,7 @@ export const PageSettings = ({
       >
         <FormFields errors={errors} values={values} onChange={handleChange} />
       </PageSettingsView>
-      {showDeleteConfirmation && page && (
-        <DeleteConfirmationDialog
-          page={page}
-          onClose={() => {
-            setShowDeleteConfirmation(false);
-          }}
-          onConfirm={() => {
-            setShowDeleteConfirmation(false);
-            hanldeDelete();
-          }}
-        />
-      )}
     </>
-  );
-};
-
-type DeleteConfirmationDialogProps = {
-  onClose: () => void;
-  onConfirm: () => void;
-  page: Page;
-};
-
-const DeleteConfirmationDialog = ({
-  onClose,
-  onConfirm,
-  page,
-}: DeleteConfirmationDialogProps) => {
-  return (
-    <Dialog
-      open
-      onOpenChange={(isOpen) => {
-        if (isOpen === false) {
-          onClose();
-        }
-      }}
-    >
-      <DialogContent>
-        <Flex gap="3" direction="column" css={{ padding: theme.panel.padding }}>
-          <Text>{`Are you sure you want to delete "${page.name}"?`}</Text>
-          <Text>
-            You can undo it even if you delete the page as long as you don't
-            reload.
-          </Text>
-          <Flex direction="rowReverse" gap="2">
-            <DialogClose>
-              <Button
-                color="destructive"
-                onClick={() => {
-                  onConfirm();
-                }}
-              >
-                Delete Page
-              </Button>
-            </DialogClose>
-            <DialogClose>
-              <Button color="ghost">Cancel</Button>
-            </DialogClose>
-          </Flex>
-        </Flex>
-        <DialogTitle>Delete Page</DialogTitle>
-      </DialogContent>
-    </Dialog>
   );
 };
